refactored extension to allow cleaner addition of stats scope
