+++ ext/config.w32
@@ -4,6 +4,6 @@
 ARG_ENABLE("opencensus", "opencensus support", "no");
 
 if (PHP_OPENCENSUS != "no") {
-    EXTENSION('opencensus', 'opencensus_trace.c opencensus_trace_context.c opencensus_trace_span.c opencensus_trace_annotation.c opencensus_trace_link.c opencensus_trace_message_event.c');
+    EXTENSION('opencensus', 'opencensus.c opencensus_trace.c opencensus_trace_context.c opencensus_trace_span.c opencensus_trace_annotation.c opencensus_trace_link.c opencensus_trace_message_event.c');
     AC_DEFINE('HAVE_OPENCENSUS', 1);
 }
