+++ ext/php_opencensus.h
@@ -30,10 +30,8 @@
 #include <sys/time.h>
 #endif
 
-#define PHP_OPENCENSUS_VERSION "0.2.2"
+#define PHP_OPENCENSUS_VERSION "0.3.0"
 #define PHP_OPENCENSUS_EXTNAME "opencensus"
-
-PHP_FUNCTION(opencensus_version);
 
 extern zend_module_entry opencensus_module_entry;
 #define phpext_opencensus_ptr &opencensus_module_entry
@@ -42,6 +40,8 @@ PHP_MINIT_FUNCTION(opencensus);
 PHP_MSHUTDOWN_FUNCTION(opencensus);
 PHP_RINIT_FUNCTION(opencensus);
 PHP_RSHUTDOWN_FUNCTION(opencensus);
+
+PHP_FUNCTION(opencensus_version);
 
 ZEND_BEGIN_MODULE_GLOBALS(opencensus)
     // map of functions we're tracing to callbacks
@@ -58,19 +58,17 @@ ZEND_END_MODULE_GLOBALS(opencensus)
 
 extern ZEND_DECLARE_MODULE_GLOBALS(opencensus)
 
-/* Return the current timestamp as a double */
-static double opencensus_now()
-{
-    struct timeval tv;
-    gettimeofday(&tv, NULL);
-
-    return (double) (tv.tv_sec + tv.tv_usec / 1000000.00);
-}
-
 #ifdef ZTS
-#define        OPENCENSUS_TRACE_G(v)        TSRMG(opencensus_globals_id, zend_opencensus_globals *, v)
+#define OPENCENSUS_G(v) TSRMG(opencensus_globals_id, zend_opencensus_globals *, v)
 #else
-#define        OPENCENSUS_TRACE_G(v)        (opencensus_globals.v)
+#define OPENCENSUS_G(v) (opencensus_globals.v)
 #endif
+
+// Extension lifecycle hooks
+int opencensus_minit(INIT_FUNC_ARGS);
+int opencensus_rinit(TSRMLS_D);
+int opencensus_rshutdown(TSRMLS_D);
+
+extern double opencensus_now();
 
 #endif /* PHP_OPENCENSUS_H */
